@@ -3,6 +3,7 @@ import { EditorConfig } from '../../types';
 import { ModuleConfig } from '../types-internal/module-config';
 import Listeners from './utils/listeners';
 import EventsDispatcher from './utils/events';
+import { EditorEventMap } from './events';
 
 /**
  * The type <T> of the Module generic.
@@ -14,12 +15,11 @@ export type ModuleNodes = object;
  * @abstract
  * @class      Module
  * @classdesc  All modules inherits from this class.
- *
  * @typedef {Module} Module
  * @property {object} config - Editor user settings
  * @property {EditorModules} Editor - List of Editor modules
  */
-export default class Module<T extends ModuleNodes = {}> {
+export default class Module<T extends ModuleNodes = Record<string, HTMLElement>> {
   /**
    * Each module can provide some UI elements that will be stored in this property
    */
@@ -43,7 +43,7 @@ export default class Module<T extends ModuleNodes = {}> {
   /**
    * Editor event dispatcher class
    */
-  protected eventsDispatcher: EventsDispatcher;
+  protected eventsDispatcher: EventsDispatcher<EditorEventMap>;
 
   /**
    * Util for bind/unbind DOM event listeners
@@ -92,8 +92,9 @@ export default class Module<T extends ModuleNodes = {}> {
 
   /**
    * @class
-   *
-   * @param {ModuleConfig} - Module config
+   * @param options - Module options
+   * @param options.config - Module config
+   * @param options.eventsDispatcher - Common event bus
    */
   constructor({ config, eventsDispatcher }: ModuleConfig) {
     if (new.target === Module) {
