@@ -1,11 +1,11 @@
 /**
  * @class MoveUpTune
  * @classdesc Editor's default tune that moves up selected block
- *
  * @copyright <CodeX Team> 2018
  */
-import $ from '../dom';
 import { API, BlockTune } from '../../../types';
+import { IconChevronUp } from '@codexteam/icons';
+import { TunesMenuConfig } from '../../../types/tools';
 
 /**
  *
@@ -25,12 +25,8 @@ export default class MoveUpTune implements BlockTune {
 
   /**
    * Styles
-   *
-   * @type {{wrapper: string}}
    */
   private CSS = {
-    button: 'ce-settings__button',
-    wrapper: 'ce-tune-move-up',
     animation: 'wobble',
   };
 
@@ -44,51 +40,27 @@ export default class MoveUpTune implements BlockTune {
   }
 
   /**
-   * Create "MoveUp" button and add click event listener
-   *
-   * @returns {HTMLElement}
+   * Tune's appearance in block settings menu
    */
-  public render(): HTMLElement {
-    const moveUpButton = $.make('div', [this.CSS.button, this.CSS.wrapper], {});
-
-    // moveUpButton.appendChild($.svg('arrow-up', 14, 14));
-    moveUpButton.innerHTML = '<svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M7.83339 12.8333V3.52493L11.9001 7.5916C12.2251 7.9166 12.7584 7.9166 13.0834 7.5916C13.4084 7.2666 13.4084 6.7416 13.0834 6.4166L7.59172 0.924933C7.26672 0.599933 6.74172 0.599933 6.41672 0.924933L0.916724 6.40827C0.591724 6.73327 0.591724 7.25827 0.916724 7.58327C1.24172 7.90827 1.76672 7.90827 2.09172 7.58327L6.16672 3.52493V12.8333C6.16672 13.2916 6.54172 13.6666 7.00006 13.6666C7.45839 13.6666 7.83339 13.2916 7.83339 12.8333Z" fill="#04003D"/></svg>';
-    this.api.listeners.on(
-      moveUpButton,
-      'click',
-      (event) => this.handleClick(event as MouseEvent, moveUpButton),
-      false
-    );
-
-    /**
-     * Enable tooltip module on button
-     */
-    this.api.tooltip.onHover(moveUpButton, this.api.i18n.t('Move up'), {
-      hidingDelay: 300,
-    });
-
-    return moveUpButton;
+  public render(): TunesMenuConfig {
+    return {
+      icon: IconChevronUp,
+      title: this.api.i18n.t('Move up'),
+      onActivate: (): void => this.handleClick(),
+      name: 'move-up',
+    };
   }
 
   /**
    * Move current block up
-   *
-   * @param {MouseEvent} event - click event
-   * @param {HTMLElement} button - clicked button
    */
-  public handleClick(event: MouseEvent, button: HTMLElement): void {
+  public handleClick(): void {
     const currentBlockIndex = this.api.blocks.getCurrentBlockIndex();
     const currentBlock = this.api.blocks.getBlockByIndex(currentBlockIndex);
     const previousBlock = this.api.blocks.getBlockByIndex(currentBlockIndex - 1);
 
     if (currentBlockIndex === 0 || !currentBlock || !previousBlock) {
-      button.classList.add(this.CSS.animation);
-
-      window.setTimeout(() => {
-        button.classList.remove(this.CSS.animation);
-      }, 500);
-
-      return;
+      throw new Error('Unable to move Block up since it is already the first');
     }
 
     const currentBlockElement = currentBlock.holder;
@@ -110,7 +82,7 @@ export default class MoveUpTune implements BlockTune {
     if (previousBlockCoords.top > 0) {
       scrollUpOffset = Math.abs(currentBlockCoords.top) - Math.abs(previousBlockCoords.top);
     } else {
-      scrollUpOffset = window.innerHeight - Math.abs(currentBlockCoords.top) + Math.abs(previousBlockCoords.top);
+      scrollUpOffset = Math.abs(currentBlockCoords.top) + previousBlockCoords.height;
     }
 
     window.scrollBy(0, -1 * scrollUpOffset);
@@ -119,8 +91,5 @@ export default class MoveUpTune implements BlockTune {
     this.api.blocks.move(currentBlockIndex - 1);
 
     this.api.toolbar.toggleBlockSettings(true);
-
-    /** Hide the Tooltip */
-    this.api.tooltip.hide();
   }
 }
