@@ -45,15 +45,14 @@ export default class Dom {
   }
 
   /**
-   * Helper for making Elements with classname and attributes
+   * Helper for making Elements with class name and attributes
    *
    * @param  {string} tagName - new Element tag name
-   * @param  {string[]|string} [classNames] - list or name of CSS classname(s)
+   * @param  {string[]|string} [classNames] - list or name of CSS class name(s)
    * @param  {object} [attributes] - any attributes
-   *
    * @returns {HTMLElement}
    */
-  public static make(tagName: string, classNames: string|string[] = null, attributes: object = {}): HTMLElement {
+  public static make(tagName: string, classNames: string | string[] = null, attributes: object = {}): HTMLElement {
     const el = document.createElement(tagName);
 
     if (Array.isArray(classNames)) {
@@ -75,31 +74,10 @@ export default class Dom {
    * Creates Text Node with the passed content
    *
    * @param {string} content - text content
-   *
    * @returns {Text}
    */
   public static text(content: string): Text {
     return document.createTextNode(content);
-  }
-
-  /**
-   * Creates SVG icon linked to the sprite
-   *
-   * @param {string} name - name (id) of icon from sprite
-   * @param {number} [width] - icon width
-   * @param {number} [height] - icon height
-   *
-   * @returns {SVGElement}
-   */
-  public static svg(name: string, width = 14, height = 14): SVGElement {
-    const icon = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
-
-    icon.classList.add('icon', 'icon--' + name);
-    icon.setAttribute('width', width + 'px');
-    icon.setAttribute('height', height + 'px');
-    icon.innerHTML = `<use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#${name}"></use>`;
-
-    return icon;
   }
 
   /**
@@ -109,8 +87,8 @@ export default class Dom {
    * @param  {Element|Element[]|DocumentFragment|Text|Text[]} elements - element or elements list
    */
   public static append(
-    parent: Element|DocumentFragment,
-    elements: Element|Element[]|DocumentFragment|Text|Text[]
+    parent: Element | DocumentFragment,
+    elements: Element | Element[] | DocumentFragment | Text | Text[]
   ): void {
     if (Array.isArray(elements)) {
       elements.forEach((el) => parent.appendChild(el));
@@ -125,7 +103,7 @@ export default class Dom {
    * @param {Element} parent - where to append
    * @param {Element|Element[]} elements - element or elements list
    */
-  public static prepend(parent: Element, elements: Element|Element[]): void {
+  public static prepend(parent: Element, elements: Element | Element[]): void {
     if (Array.isArray(elements)) {
       elements = elements.reverse();
       elements.forEach((el) => parent.prepend(el));
@@ -165,10 +143,9 @@ export default class Dom {
    *
    * @param {Element} el - element we searching inside. Default - DOM Document
    * @param {string} selector - searching string
-   *
    * @returns {Element}
    */
-  public static find(el: Element|Document = document, selector: string): Element {
+  public static find(el: Element | Document = document, selector: string): Element | null {
     return el.querySelector(selector);
   }
 
@@ -189,10 +166,9 @@ export default class Dom {
    *
    * @param {Element|Document} el - element we searching inside. Default - DOM Document
    * @param {string} selector - searching string
-   *
    * @returns {NodeList}
    */
-  public static findAll(el: Element|Document = document, selector: string): NodeList {
+  public static findAll(el: Element | Document = document, selector: string): NodeList {
     return el.querySelectorAll(selector);
   }
 
@@ -207,7 +183,7 @@ export default class Dom {
   }
 
   /**
-   * Find all contendeditable, textarea and editable input elements passed holder contains
+   * Find all contenteditable, textarea and editable input elements passed holder contains
    *
    * @param holder - element where to find inputs
    */
@@ -230,11 +206,9 @@ export default class Dom {
    * Leaf is the vertex that doesn't have any child nodes
    *
    * @description Method recursively goes throw the all Node until it finds the Leaf
-   *
    * @param {Node} node - root Node. From this vertex we start Deep-first search
    *                      {@link https://en.wikipedia.org/wiki/Depth-first_search}
    * @param {boolean} [atLast] - find last text node
-   *
    * @returns {Node} - it can be text Node or Element Node, so that caret will able to work with it
    */
   public static getDeepestNode(node: Node, atLast = false): Node {
@@ -287,7 +261,6 @@ export default class Dom {
    * Check if object is DOM node
    *
    * @param {*} node - object to check
-   *
    * @returns {boolean}
    */
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -318,7 +291,6 @@ export default class Dom {
    * Check if passed element is contenteditable
    *
    * @param {HTMLElement} element - html element to check
-   *
    * @returns {boolean}
    */
   public static isContentEditable(element: HTMLElement): boolean {
@@ -329,7 +301,6 @@ export default class Dom {
    * Checks target if it is native input
    *
    * @param {*} target - HTML element or string
-   *
    * @returns {boolean}
    */
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -346,7 +317,6 @@ export default class Dom {
    * Checks if we can set caret
    *
    * @param {HTMLElement} target - target to check
-   *
    * @returns {boolean}
    */
   public static canSetCaret(target: HTMLElement): boolean {
@@ -377,9 +347,7 @@ export default class Dom {
    *
    * @description Method checks simple Node without any childs for emptiness
    * If you have Node with 2 or more children id depth, you better use {@link Dom#isEmpty} method
-   *
    * @param {Node} node - node to check
-   *
    * @returns {boolean} true if it is empty
    */
   public static isNodeEmpty(node: Node): boolean {
@@ -402,7 +370,6 @@ export default class Dom {
    * checks node if it is doesn't have any child nodes
    *
    * @param {Node} node - node to check
-   *
    * @returns {boolean}
    */
   public static isLeaf(node: Node): boolean {
@@ -418,7 +385,6 @@ export default class Dom {
    * {@link https://en.wikipedia.org/wiki/Breadth-first_search}
    *
    * @description Pushes to stack all DOM leafs and checks for emptiness
-   *
    * @param {Node} node - node to check
    * @returns {boolean}
    */
@@ -453,7 +419,6 @@ export default class Dom {
    * Check if string contains html elements
    *
    * @param {string} str - string to check
-   *
    * @returns {boolean}
    */
   public static isHTMLString(str: string): boolean {
@@ -468,7 +433,6 @@ export default class Dom {
    * Return length of node`s text content
    *
    * @param {Node} node - node with content
-   *
    * @returns {number}
    */
   public static getContentLength(node: Node): number {
@@ -523,6 +487,8 @@ export default class Dom {
       'ruby',
       'section',
       'table',
+      'tbody',
+      'thead',
       'tr',
       'tfoot',
       'ul',
@@ -534,7 +500,6 @@ export default class Dom {
    * Check if passed content includes only inline elements
    *
    * @param {string|HTMLElement} data - element or html string
-   *
    * @returns {boolean}
    */
   public static containsOnlyInlineElements(data: string | HTMLElement): boolean {
@@ -559,7 +524,6 @@ export default class Dom {
    * Find and return all block elements in the passed parent (including subtree)
    *
    * @param {HTMLElement} parent - root element
-   *
    * @returns {HTMLElement[]}
    */
   public static getDeepestBlockElements(parent: HTMLElement): HTMLElement[] {
@@ -576,7 +540,6 @@ export default class Dom {
    * Helper for get holder from {string} or return HTMLElement
    *
    * @param {string | HTMLElement} element - holder's id or holder's HTML Element
-   *
    * @returns {HTMLElement}
    */
   public static getHolder(element: string | HTMLElement): HTMLElement {
@@ -588,25 +551,9 @@ export default class Dom {
   }
 
   /**
-   * Method checks passed Node if it is some extension Node
-   *
-   * @param {Node} node - any node
-   *
-   * @returns {boolean}
-   */
-  public static isExtensionNode(node: Node): boolean {
-    const extensions = [
-      'GRAMMARLY-EXTENSION',
-    ];
-
-    return node && extensions.includes(node.nodeName);
-  }
-
-  /**
    * Returns true if element is anchor (is A tag)
    *
    * @param {Element} element - element to check
-   *
    * @returns {boolean}
    */
   public static isAnchor(element: Element): element is HTMLAnchorElement {
@@ -619,7 +566,7 @@ export default class Dom {
    * @todo handle case when editor initialized in scrollable popup
    * @param el - element to compute offset
    */
-  public static offset(el): {top: number; left: number; right: number; bottom: number} {
+  public static offset(el): { top: number; left: number; right: number; bottom: number } {
     const rect = el.getBoundingClientRect();
     const scrollLeft = window.pageXOffset || document.documentElement.scrollLeft;
     const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
