@@ -1,5 +1,5 @@
-import $ from '../dom';
 import { InlineTool, SanitizerConfig } from '../../../types';
+import { IconBold } from '@codexteam/icons';
 
 /**
  * Bold Tool
@@ -61,28 +61,25 @@ export default class BoldInlineTool implements InlineTool {
     this.nodes.button = document.createElement('button') as HTMLButtonElement;
     this.nodes.button.type = 'button';
     this.nodes.button.classList.add(this.CSS.button, this.CSS.buttonModifier);
-    this.nodes.button.appendChild($.svg('bold', 24, 24));
+    // TODO fix icon
+    this.nodes.button.innerHTML = IconBold;
 
     return this.nodes.button;
   }
 
   /**
    * Wrap range with <b> tag
-   *
-   * @param {Range} range - range to wrap
    */
-  public surround(range: Range): void {
+  public surround(): void {
     document.execCommand(this.commandName);
   }
 
   /**
    * Check selection and set activated state to button if there are <b> tag
    *
-   * @param {Selection} selection - selection to check
-   *
    * @returns {boolean}
    */
-  public checkState(selection: Selection): boolean {
+  public checkState(): boolean {
     const isActive = document.queryCommandState(this.commandName);
 
     this.nodes.button.classList.toggle(this.CSS.buttonActive, isActive);
