@@ -27,7 +27,6 @@ export default class CaretAPI extends Module {
    *
    * @param {string} position - position where to set caret
    * @param {number} offset - caret offset
-   *
    * @returns {boolean}
    */
   private setToFirstBlock = (position: string = this.Editor.Caret.positions.DEFAULT, offset = 0): boolean => {
@@ -38,14 +37,13 @@ export default class CaretAPI extends Module {
     this.Editor.Caret.setToBlock(this.Editor.BlockManager.firstBlock, position, offset);
 
     return true;
-  }
+  };
 
   /**
    * Sets caret to the last Block
    *
    * @param {string} position - position where to set caret
    * @param {number} offset - caret offset
-   *
    * @returns {boolean}
    */
   private setToLastBlock = (position: string = this.Editor.Caret.positions.DEFAULT, offset = 0): boolean => {
@@ -56,14 +54,13 @@ export default class CaretAPI extends Module {
     this.Editor.Caret.setToBlock(this.Editor.BlockManager.lastBlock, position, offset);
 
     return true;
-  }
+  };
 
   /**
    * Sets caret to the previous Block
    *
    * @param {string} position - position where to set caret
    * @param {number} offset - caret offset
-   *
    * @returns {boolean}
    */
   private setToPreviousBlock = (
@@ -77,14 +74,13 @@ export default class CaretAPI extends Module {
     this.Editor.Caret.setToBlock(this.Editor.BlockManager.previousBlock, position, offset);
 
     return true;
-  }
+  };
 
   /**
    * Sets caret to the next Block
    *
    * @param {string} position - position where to set caret
    * @param {number} offset - caret offset
-   *
    * @returns {boolean}
    */
   private setToNextBlock = (position: string = this.Editor.Caret.positions.DEFAULT, offset = 0): boolean => {
@@ -95,7 +91,7 @@ export default class CaretAPI extends Module {
     this.Editor.Caret.setToBlock(this.Editor.BlockManager.nextBlock, position, offset);
 
     return true;
-  }
+  };
 
   /**
    * Sets caret to the Block by passed index
@@ -103,7 +99,6 @@ export default class CaretAPI extends Module {
    * @param {number} index - index of Block where to set caret
    * @param {string} position - position where to set caret
    * @param {number} offset - caret offset
-   *
    * @returns {boolean}
    */
   private setToBlock = (
@@ -118,13 +113,12 @@ export default class CaretAPI extends Module {
     this.Editor.Caret.setToBlock(this.Editor.BlockManager.blocks[index], position, offset);
 
     return true;
-  }
+  };
 
   /**
    * Sets caret to the Editor
    *
    * @param {boolean} atEnd - if true, set Caret to the end of the Editor
-   *
    * @returns {boolean}
    */
   private focus = (atEnd = false): boolean => {
@@ -133,5 +127,5 @@ export default class CaretAPI extends Module {
     }
 
     return this.setToFirstBlock(this.Editor.Caret.positions.START);
-  }
+  };
 }
