@@ -1,29 +1,23 @@
 import Module from '../../__module';
 import $ from '../../dom';
-import Flipper, { FlipperOptions } from '../../flipper';
-import * as _ from '../../utils';
 import SelectionUtils from '../../selection';
 import Block from '../../block';
+import I18n from '../../i18n';
+import { I18nInternalNS } from '../../i18n/namespace-internal';
+import Flipper from '../../flipper';
+import { TunesMenuConfigItem } from '../../../../types/tools';
+import { resolveAliases } from '../../utils/resolve-aliases';
+import Popover, { PopoverEvent } from '../../utils/popover';
 
 /**
  * HTML Elements that used for BlockSettings
  */
 interface BlockSettingsNodes {
   wrapper: HTMLElement;
-  toolSettings: HTMLElement;
-  defaultSettings: HTMLElement;
 }
 
 /**
  * Block Settings
- *
- *   ____ Settings Panel ____
- *  | ...................... |
- *  | .   Tool Settings    . |
- *  | ...................... |
- *  | .  Default Settings  . |
- *  | ...................... |
- *  |________________________|
  *
  *  @todo Make Block Settings no-module but a standalone class, like Toolbox
  */
@@ -42,44 +36,26 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
 
   /**
    * Block Settings CSS
-   *
-   * @returns {{wrapper, wrapperOpened, toolSettings, defaultSettings, button}}
    */
   public get CSS(): { [name: string]: string } {
     return {
-      // Settings Panel
-      wrapper: 'ce-settings',
-      wrapperOpened: 'ce-settings--opened',
-      toolSettings: 'ce-settings__plugin-zone',
-      defaultSettings: 'ce-settings__default-zone',
-
-      button: 'ce-settings__button',
-
-      focusedButton: 'ce-settings__button--focused',
-      focusedButtonAnimated: 'ce-settings__button--focused-animated',
+      settings: 'ce-settings',
     };
   }
 
   /**
-   * Is Block Settings opened or not
-   *
-   * @returns {boolean}
+   * Opened state
    */
-  public get opened(): boolean {
-    return this.nodes.wrapper.classList.contains(this.CSS.wrapperOpened);
+  public opened = false;
+
+  /**
+   * Getter for inner popover's flipper instance
+   *
+   * @todo remove once BlockSettings becomes standalone non-module class
+   */
+  public get flipper(): Flipper {
+    return this.popover?.flipper;
   }
-
-  /**
-   * List of buttons
-   */
-  private buttons: HTMLElement[] = [];
-
-  /**
-   * Instance of class that responses for leafing buttons by arrows/tab
-   *
-   * @type {Flipper|null}
-   */
-  private flipper: Flipper = null;
 
   /**
    * Page selection utils
@@ -87,37 +63,24 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
   private selection: SelectionUtils = new SelectionUtils();
 
   /**
+   * Popover instance. There is a util for vertical lists.
+   */
+  private popover: Popover | undefined;
+
+
+  /**
    * Panel with block settings with 2 sections:
    *  - Tool's Settings
    *  - Default Settings [Move, Remove, etc]
    */
   public make(): void {
-    this.nodes.wrapper = $.make('div', this.CSS.wrapper);
-
-    this.nodes.toolSettings = $.make('div', this.CSS.toolSettings);
-    this.nodes.defaultSettings = $.make('div', this.CSS.defaultSettings);
-
-    $.append(this.nodes.wrapper, [this.nodes.toolSettings, this.nodes.defaultSettings]);
-
-    /**
-     * Active leafing by arrows/tab
-     * Buttons will be filled on opening
-     */
-    this.enableFlipper();
+    this.nodes.wrapper = $.make('div', [ this.CSS.settings ]);
   }
 
   /**
    * Destroys module
    */
   public destroy(): void {
-    /**
-     * Sometimes (in read-only mode) there is no Flipper
-     */
-    if (this.flipper) {
-      this.flipper.deactivate();
-      this.flipper = null;
-    }
-
     this.removeAllNodes();
   }
 
@@ -127,7 +90,7 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
    * @param targetBlock - near which Block we should open BlockSettings
    */
   public open(targetBlock: Block = this.Editor.BlockManager.currentBlock): void {
-    this.nodes.wrapper.classList.add(this.CSS.wrapperOpened);
+    this.opened = true;
 
     /**
      * If block settings contains any inputs, focus will be set there,
@@ -144,24 +107,41 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
     /**
      * Fill Tool's settings
      */
-    this.addToolSettings(targetBlock);
-
-    /**
-     * Add default settings that presents for all Blocks
-     */
-    this.addTunes(targetBlock);
+    const [tunesItems, customHtmlTunesContainer] = targetBlock.getTunes();
 
     /** Tell to subscribers that block settings is opened */
     this.eventsDispatcher.emit(this.events.opened);
+    this.popover = new Popover({
+      searchable: true,
+      items: tunesItems.map(tune => this.resolveTuneAliases(tune)),
+      customContent: customHtmlTunesContainer,
+      customContentFlippableItems: this.getControls(customHtmlTunesContainer),
+      scopeElement: this.Editor.API.methods.ui.nodes.redactor,
+      messages: {
+        nothingFound: I18n.ui(I18nInternalNS.ui.popover, 'Nothing found'),
+        search: I18n.ui(I18nInternalNS.ui.popover, 'Filter'),
+      },
+    });
 
-    this.flipper.activate(this.blockTunesButtons);
+    this.popover.on(PopoverEvent.Close, this.onPopoverClose);
+
+    this.nodes.wrapper.append(this.popover.getElement());
+
+    this.popover.show();
+  }
+
+  /**
+   * Returns root block settings element
+   */
+  public getElement(): HTMLElement {
+    return this.nodes.wrapper;
   }
 
   /**
    * Close Block Settings pane
    */
   public close(): void {
-    this.nodes.wrapper.classList.remove(this.CSS.wrapperOpened);
+    this.opened = false;
 
     /**
      * If selection is at editor on Block Settings closing,
@@ -183,106 +163,51 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
       this.Editor.BlockManager.currentBlock.selected = false;
     }
 
-    /** Clear settings */
-    this.nodes.toolSettings.innerHTML = '';
-    this.nodes.defaultSettings.innerHTML = '';
-
     /** Tell to subscribers that block settings is closed */
     this.eventsDispatcher.emit(this.events.closed);
 
-    /** Clear cached buttons */
-    this.buttons = [];
-
-    /** Clear focus on active button */
-    this.flipper.deactivate();
+    if (this.popover) {
+      this.popover.off(PopoverEvent.Close, this.onPopoverClose);
+      this.popover.destroy();
+      this.popover.getElement().remove();
+      this.popover = null;
+    }
   }
 
   /**
-   * Returns Tools Settings and Default Settings
-   *
-   * @returns {HTMLElement[]}
+   * Handles popover close event
    */
-  public get blockTunesButtons(): HTMLElement[] {
+  private onPopoverClose = (): void => {
+    this.close();
+  };
+
+  /**
+   * Returns list of buttons and inputs inside specified container
+   *
+   * @param container - container to query controls inside of
+   */
+  private getControls(container: HTMLElement): HTMLElement[] {
     const { StylesAPI } = this.Editor;
-
-    /**
-     * Return from cache
-     * if exists
-     */
-    if (this.buttons.length !== 0) {
-      return this.buttons;
-    }
-
-    const toolSettings = this.nodes.toolSettings.querySelectorAll(
-      // Select buttons and inputs
+    /** Query buttons and inputs inside tunes html */
+    const controls = container.querySelectorAll<HTMLElement>(
       `.${StylesAPI.classes.settingsButton}, ${$.allInputsSelector}`
     );
-    const defaultSettings = this.nodes.defaultSettings.querySelectorAll(`.${this.CSS.button}`);
 
-    toolSettings.forEach((item) => {
-      this.buttons.push((item as HTMLElement));
-    });
-
-    defaultSettings.forEach((item) => {
-      this.buttons.push((item as HTMLElement));
-    });
-
-    return this.buttons;
+    return Array.from(controls);
   }
 
   /**
-   * Add Tool's settings
+   * Resolves aliases in tunes menu items
    *
-   * @param targetBlock - Block to render settings
+   * @param item - item with resolved aliases
    */
-  private addToolSettings(targetBlock): void {
-    const settingsElement = targetBlock.renderSettings();
+  private resolveTuneAliases(item: TunesMenuConfigItem): TunesMenuConfigItem {
+    const result = resolveAliases(item, { label: 'title' });
 
-    if (settingsElement) {
-      $.append(this.nodes.toolSettings, settingsElement);
+    if (item.confirmation) {
+      result.confirmation = this.resolveTuneAliases(item.confirmation);
     }
-  }
 
-  /**
-   * Add tunes: provided by user and default ones
-   *
-   * @param targetBlock - Block to render its Tunes set
-   */
-  private addTunes(targetBlock): void {
-    const [toolTunes, defaultTunes] = targetBlock.renderTunes();
-
-    $.append(this.nodes.toolSettings, toolTunes);
-    $.append(this.nodes.defaultSettings, defaultTunes);
-  }
-
-  /**
-   * Active leafing by arrows/tab
-   * Buttons will be filled on opening
-   */
-  private enableFlipper(): void {
-    this.flipper = new Flipper({
-      focusedItemClass: this.CSS.focusedButton,
-      /**
-       * @param {HTMLElement} focusedItem - activated Tune
-       */
-      activateCallback: (focusedItem) => {
-        /**
-         * If focused item is editable element, close block settings
-         */
-        if (focusedItem && $.canSetCaret(focusedItem)) {
-          this.close();
-
-          return;
-        }
-
-        /**
-         * Restoring focus on current Block after settings clicked.
-         * For example, when H3 changed to H2 — DOM Elements replaced, so we need to focus a new one
-         */
-        _.delay(() => {
-          this.Editor.Caret.setToBlock(this.Editor.BlockManager.currentBlock);
-        }, 50)();
-      },
-    } as FlipperOptions);
+    return result;
   }
 }
