@@ -11,6 +11,7 @@ import Tooltip from '../../utils/tooltip';
 import { ModuleConfig } from '../../../types-internal/module-config';
 import InlineTool from '../../tools/inline';
 import { CommonInternalSettings } from '../../tools/base';
+import { IconChevronDown } from '@codexteam/icons';
 
 /**
  * Inline Toolbar elements
@@ -53,6 +54,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     inputField: 'cdx-input',
     focusedButton: 'ce-inline-tool--focused',
     conversionToggler: 'ce-inline-toolbar__dropdown',
+    conversionTogglerArrow: 'ce-inline-toolbar__dropdown-arrow',
     conversionTogglerHidden: 'ce-inline-toolbar__dropdown--hidden',
     conversionTogglerContent: 'ce-inline-toolbar__dropdown-content',
     conversionTogglerIcon: 'ce-inline-toolbar__dropdown-icon',
@@ -70,7 +72,8 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
   /**
    * Margin above/below the Toolbar
    */
-  private readonly toolbarVerticalMargin: number = 5;
+  // eslint-disable-next-line @typescript-eslint/no-magic-numbers
+  private readonly toolbarVerticalMargin: number = _.isMobileScreen() ? 20 : 6;
 
   /**
    * TODO: Get rid of this
@@ -123,7 +126,9 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    */
   public toggleReadOnly(readOnlyEnabled: boolean): void {
     if (!readOnlyEnabled) {
-      this.make();
+      window.requestIdleCallback(() => {
+        this.make();
+      }, { timeout: 2000 });
     } else {
       this.destroy();
       this.Editor.ConversionToolbar.destroy();
@@ -284,7 +289,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
   /**
    * Check if node is contained by Inline Toolbar
    *
-   * @param {Node} node — node to chcek
+   * @param {Node} node — node to check
    */
   public containsNode(node: Node): boolean {
     return this.nodes.wrapper.contains(node);
@@ -326,7 +331,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       const isClickedOnActionsWrapper = (event.target as Element).closest(`.${this.CSS.actionsWrapper}`);
 
       // If click is on actions wrapper,
-      // do not prevent default behaviour because actions might include interactive elements
+      // do not prevent default behavior because actions might include interactive elements
       if (!isClickedOnActionsWrapper) {
         event.preventDefault();
       }
@@ -360,8 +365,11 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
 
     /**
      * Recalculate initial width with all buttons
+     * We use RIC to prevent forced layout during editor initialization to make it faster
      */
-    this.recalculateWidth();
+    window.requestAnimationFrame(() => {
+      this.recalculateWidth();
+    });
 
     /**
      * Allow to leaf buttons by arrows / tab
@@ -434,12 +442,14 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     this.nodes.conversionTogglerIcon = $.make('div', this.CSS.conversionTogglerIcon);
     this.nodes.conversionTogglerText = $.make('div', this.CSS.conversionTogglerText);
 
-    const icon = $.svg('toggler-down', 13, 13);
+    const iconWrapper = $.make('div', this.CSS.conversionTogglerArrow, {
+      innerHTML: IconChevronDown,
+    });
 
     this.nodes.conversionTogglerContent.appendChild(this.nodes.conversionTogglerIcon);
     this.nodes.conversionTogglerContent.appendChild(this.nodes.conversionTogglerText);
     this.nodes.conversionToggler.appendChild(this.nodes.conversionTogglerContent);
-    this.nodes.conversionToggler.appendChild(icon);
+    this.nodes.conversionToggler.appendChild(iconWrapper);
 
     this.nodes.togglerAndButtonsWrapper.appendChild(this.nodes.conversionToggler);
 
@@ -462,10 +472,12 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       });
     });
 
-    this.tooltip.onHover(this.nodes.conversionToggler, I18n.ui(I18nInternalNS.ui.inlineToolbar.converter, 'Convert to'), {
-      placement: 'top',
-      hidingDelay: 100,
-    });
+    if (_.isMobileScreen() === false ) {
+      this.tooltip.onHover(this.nodes.conversionToggler, I18n.ui(I18nInternalNS.ui.inlineToolbar.converter, 'Convert to'), {
+        placement: 'top',
+        hidingDelay: 100,
+      });
+    }
   }
 
   /**
@@ -589,10 +601,12 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       }));
     }
 
-    this.tooltip.onHover(button, tooltipContent, {
-      placement: 'top',
-      hidingDelay: 100,
-    });
+    if (_.isMobileScreen() === false ) {
+      this.tooltip.onHover(button, tooltipContent, {
+        placement: 'top',
+        hidingDelay: 100,
+      });
+    }
 
     instance.checkState(SelectionUtils.get());
   }
@@ -672,6 +686,15 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
 
     tool.surround(range);
     this.checkToolsState();
+
+    /**
+     * If tool has "actions", so after click it will probably toggle them on.
+     * For example, the Inline Link Tool will show the URL-input.
+     * So we disable the Flipper for that case to allow Tool bind own Enter listener
+     */
+    if (tool.renderActions !== undefined) {
+      this.flipper.deactivate();
+    }
   }
 
   /**
