@@ -1,4 +1,4 @@
-import Paragraph from '../../tools/paragraph/dist/bundle';
+import Paragraph from '@editorjs/paragraph';
 import Module from '../__module';
 import * as _ from '../utils';
 import { SanitizerConfig, ToolConfig, ToolConstructable, ToolSettings } from '../../../types';
@@ -23,17 +23,8 @@ import ToolsCollection from '../tools/collection';
  * Creates Instances from Plugins and binds external config to the instances
  */
 
-type ToolClass = BlockTool | InlineTool | BlockTune;
-
 /**
- * Class properties:
- *
- * @typedef {Tools} Tools
- * @property {Tools[]} toolsAvailable - available Tools
- * @property {Tools[]} toolsUnavailable - unavailable Tools
- * @property {object} toolsClasses - all classes
- * @property {object} toolsSettings - Tools settings
- * @property {EditorConfig} config - Editor config
+ * Modules that works with tools classes
  */
 export default class Tools extends Module {
   /**
@@ -46,8 +37,6 @@ export default class Tools extends Module {
 
   /**
    * Returns available Tools
-   *
-   * @returns {object<Tool>}
    */
   public get available(): ToolsCollection {
     return this.toolsAvailable;
@@ -55,8 +44,6 @@ export default class Tools extends Module {
 
   /**
    * Returns unavailable Tools
-   *
-   * @returns {Tool[]}
    */
   public get unavailable(): ToolsCollection {
     return this.toolsUnavailable;
@@ -64,8 +51,6 @@ export default class Tools extends Module {
 
   /**
    * Return Tools for the Inline Toolbar
-   *
-   * @returns {object} - object of Inline Tool's classes
    */
   public get inlineTools(): ToolsCollection<InlineTool> {
     return this.available.inlineTools;
