@@ -146,7 +146,7 @@ export default class SelectionUtils {
   /**
    * Check if passed selection is at Editor's zone
    *
-   * @param selection - Selectoin object to check
+   * @param selection - Selection object to check
    */
   public static isSelectionAtEditor(selection: Selection): boolean {
     if (!selection) {
@@ -225,7 +225,7 @@ export default class SelectionUtils {
    *
    * @param selection - Selection object to get Range from
    */
-  public static getRangeFromSelection(selection: Selection): Range {
+  public static getRangeFromSelection(selection: Selection): Range | null {
     return selection && selection.rangeCount ? selection.getRangeAt(0) : null;
   }
 
@@ -326,8 +326,6 @@ export default class SelectionUtils {
    *
    * @param element - element where to set focus
    * @param offset - offset of cursor
-   *
-   * @returns {DOMRect} of range
    */
   public static setCursor(element: HTMLElement, offset = 0): DOMRect {
     const range = document.createRange();
@@ -355,22 +353,45 @@ export default class SelectionUtils {
   }
 
   /**
-   * Adds fake cursor to the current range
+   * Check if current range exists and belongs to container
    *
-   * @param [container] - if passed cursor will be added only if container contains current range
+   * @param container - where range should be
    */
-  public static addFakeCursor(container?: HTMLElement): void {
+  public static isRangeInsideContainer(container: HTMLElement): boolean {
     const range = SelectionUtils.range;
+
+    if (range === null) {
+      return false;
+    }
+
+    return container.contains(range.startContainer);
+  }
+
+  /**
+   * Adds fake cursor to the current range
+   */
+  public static addFakeCursor(): void {
+    const range = SelectionUtils.range;
+
+    if (range === null) {
+      return;
+    }
+
     const fakeCursor = $.make('span', 'codex-editor__fake-cursor');
 
     fakeCursor.dataset.mutationFree = 'true';
 
-    if (!range || (container && !container.contains(range.startContainer))) {
-      return;
-    }
-
     range.collapse();
     range.insertNode(fakeCursor);
+  }
+
+  /**
+   * Check if passed element contains a fake cursor
+   *
+   * @param el - where to check
+   */
+  public static isFakeCursorInsideContainer(el: HTMLElement): boolean {
+    return $.find(el, `.codex-editor__fake-cursor`) !== null;
   }
 
   /**
@@ -381,7 +402,11 @@ export default class SelectionUtils {
   public static removeFakeCursor(container: HTMLElement = document.body): void {
     const fakeCursor = $.find(container, `.codex-editor__fake-cursor`);
 
-    fakeCursor && fakeCursor.remove();
+    if (!fakeCursor) {
+      return;
+    }
+
+    fakeCursor.remove();
   }
 
   /**
@@ -452,7 +477,6 @@ export default class SelectionUtils {
    * @param  {string} tagName       - tag to found
    * @param  {string} [className]   - tag's class name
    * @param  {number} [searchDepth] - count of tags that can be included. For better performance.
-   *
    * @returns {HTMLElement|null}
    */
   public findParentTag(tagName: string, className?: string, searchDepth = 10): HTMLElement | null {
@@ -526,7 +550,7 @@ export default class SelectionUtils {
   /**
    * Expands selection range to the passed parent node
    *
-   * @param {HTMLElement} element - element which contents should be selcted
+   * @param {HTMLElement} element - element which contents should be selected
    */
   public expandToTag(element: HTMLElement): void {
     const selection = window.getSelection();
