@@ -3,9 +3,9 @@ import { BlockToolAPI } from '../block';
 import Shortcuts from '../utils/shortcuts';
 import BlockTool from '../tools/block';
 import ToolsCollection from '../tools/collection';
-import { API, BlockToolData, ToolboxConfigEntry } from '../../../types';
+import { API, BlockAPI, BlockToolData, PopoverItem, ToolboxConfigEntry } from '../../../types';
 import EventsDispatcher from '../utils/events';
-import Popover, { PopoverEvent, PopoverItem } from '../utils/popover';
+import Popover, { PopoverEvent } from '../utils/popover';
 import I18n from '../i18n';
 import { I18nInternalNS } from '../i18n/namespace-internal';
 
@@ -34,9 +34,22 @@ export enum ToolboxEvent {
 }
 
 /**
+ * Events fired by the Toolbox
+ *
+ * Event name -> payload
+ */
+export interface ToolboxEventMap {
+  [ToolboxEvent.Opened]: undefined;
+  [ToolboxEvent.Closed]: undefined;
+  [ToolboxEvent.BlockAdded]: {
+    block: BlockAPI
+  };
+}
+
+/**
  * Available i18n dict keys that should be passed to the constructor
  */
-type toolboxTextLabelsKeys = 'filter' | 'nothingFound';
+type ToolboxTextLabelsKeys = 'filter' | 'nothingFound';
 
 /**
  * Toolbox
@@ -45,7 +58,7 @@ type toolboxTextLabelsKeys = 'filter' | 'nothingFound';
  *
  * @implements {EventsDispatcher} with some events, see {@link ToolboxEvent}
  */
-export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
+export default class Toolbox extends EventsDispatcher<ToolboxEventMap> {
   /**
    * Returns True if Toolbox is Empty and nothing to show
    *
@@ -70,7 +83,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
   /**
    * Popover instance. There is a util for vertical lists.
    */
-  private popover: Popover;
+  private popover: Popover | undefined;
 
   /**
    * List of Tools available. Some of them will be shown in the Toolbox
@@ -80,33 +93,27 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
   /**
    * Text labels used in the Toolbox. Should be passed from the i18n module
    */
-  private i18nLabels: Record<toolboxTextLabelsKeys, string>;
+  private i18nLabels: Record<ToolboxTextLabelsKeys, string>;
 
   /**
    * Current module HTML Elements
    */
   private nodes: {
-    toolbox: HTMLElement;
+    toolbox: HTMLElement | null;
   } = {
-    toolbox: null,
-  };
+      toolbox: null,
+    };
 
   /**
    * CSS styles
    *
-   * @returns {object.<string, string>}
+   * @returns {Object<string, string>}
    */
   private static get CSS(): { [name: string]: string } {
     return {
       toolbox: 'ce-toolbox',
-      toolboxOpenedTop: 'ce-toolbox--opened-top',
     };
   }
-
-  /**
-   * Id of listener added used to remove it on destroy()
-   */
-  private clickListenerId: string = null;
 
   /**
    * Toolbox constructor
@@ -115,7 +122,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
    * @param options.api - Editor API methods
    * @param options.tools - Tools available to check whether some of them should be displayed at the Toolbox or not
    */
-  constructor({ api, tools, i18nLabels }: {api: API; tools: ToolsCollection<BlockTool>; i18nLabels: Record<toolboxTextLabelsKeys, string>}) {
+  constructor({ api, tools, i18nLabels }: {api: API; tools: ToolsCollection<BlockTool>; i18nLabels: Record<ToolboxTextLabelsKeys, string>}) {
     super();
 
     this.api = api;
@@ -128,14 +135,16 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
    */
   public make(): Element {
     this.popover = new Popover({
-      className: Toolbox.CSS.toolbox,
+      scopeElement: this.api.ui.nodes.redactor,
       searchable: false,
-      filterLabel: this.i18nLabels.filter,
-      nothingFoundLabel: this.i18nLabels.nothingFound,
+      messages: {
+        nothingFound: this.i18nLabels.nothingFound,
+        search: this.i18nLabels.filter,
+      },
       items: this.toolboxItemsToBeDisplayed,
     });
 
-    this.popover.on(PopoverEvent.OverlayClicked, this.onOverlayClicked);
+    this.popover.on(PopoverEvent.Close, this.onPopoverClose);
 
     /**
      * Enable tools shortcuts
@@ -143,6 +152,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
     this.enableShortcuts();
 
     this.nodes.toolbox = this.popover.getElement();
+    this.nodes.toolbox.classList.add(Toolbox.CSS.toolbox);
 
     return this.nodes.toolbox;
   }
@@ -150,8 +160,8 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
   /**
    * Returns true if the Toolbox has the Flipper activated and the Flipper has selected button
    */
-  public hasFocus(): boolean {
-    return this.popover.hasFocus();
+  public hasFocus(): boolean | undefined {
+    return this.popover?.hasFocus();
   }
 
   /**
@@ -165,10 +175,8 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
       this.nodes.toolbox = null;
     }
 
-    this.api.listeners.offById(this.clickListenerId);
-
     this.removeAllShortcuts();
-    this.popover.off(PopoverEvent.OverlayClicked, this.onOverlayClicked);
+    this.popover?.off(PopoverEvent.Close, this.onPopoverClose);
   }
 
   /**
@@ -189,16 +197,7 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
       return;
     }
 
-    /**
-     * Open the popover above the button
-     * if there is not enough available space below it
-     */
-    if (!this.shouldOpenPopoverBottom) {
-      this.nodes.toolbox.style.setProperty('--popover-height', this.popover.calculateHeight() + 'px');
-      this.nodes.toolbox.classList.add(Toolbox.CSS.toolboxOpenedTop);
-    }
-
-    this.popover.show();
+    this.popover?.show();
     this.opened = true;
     this.emit(ToolboxEvent.Opened);
   }
@@ -207,9 +206,8 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
    * Close Toolbox
    */
   public close(): void {
-    this.popover.hide();
+    this.popover?.hide();
     this.opened = false;
-    this.nodes.toolbox.classList.remove(Toolbox.CSS.toolboxOpenedTop);
     this.emit(ToolboxEvent.Closed);
   }
 
@@ -225,50 +223,29 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
   }
 
   /**
-   * Checks if there popover should be opened downwards.
-   * It happens in case there is enough space below or not enough space above
+   * Handles popover close event
    */
-  private get shouldOpenPopoverBottom(): boolean {
-    const toolboxRect = this.nodes.toolbox.getBoundingClientRect();
-    const editorElementRect = this.api.ui.nodes.redactor.getBoundingClientRect();
-    const popoverHeight = this.popover.calculateHeight();
-    const popoverPotentialBottomEdge = toolboxRect.top + popoverHeight;
-    const popoverPotentialTopEdge = toolboxRect.top - popoverHeight;
-    const bottomEdgeForComparison = Math.min(window.innerHeight, editorElementRect.bottom);
-
-    return popoverPotentialTopEdge < editorElementRect.top || popoverPotentialBottomEdge <= bottomEdgeForComparison;
-  }
-
-  /**
-   * Handles overlay click
-   */
-  private onOverlayClicked = (): void => {
-    this.close();
-  }
+  private onPopoverClose = (): void => {
+    this.opened = false;
+    this.emit(ToolboxEvent.Closed);
+  };
 
   /**
    * Returns list of tools that enables the Toolbox (by specifying the 'toolbox' getter)
    */
   @_.cacheable
   private get toolsToBeDisplayed(): BlockTool[] {
-    return Array
-      .from(this.tools.values())
-      .reduce((result, tool) => {
-        const toolToolboxSettings = tool.toolbox;
+    const result: BlockTool[] = [];
 
-        if (toolToolboxSettings) {
-          const validToolboxSettings = toolToolboxSettings.filter(item => {
-            return this.areToolboxSettingsValid(item, tool.name);
-          });
+    this.tools.forEach((tool) => {
+      const toolToolboxSettings = tool.toolbox;
 
-          result.push({
-            ...tool,
-            toolbox: validToolboxSettings,
-          });
-        }
+      if (toolToolboxSettings) {
+        result.push(tool);
+      }
+    });
 
-        return result;
-      }, []);
+    return result;
   }
 
   /**
@@ -282,9 +259,9 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
     const toPopoverItem = (toolboxItem: ToolboxConfigEntry, tool: BlockTool): PopoverItem => {
       return {
         icon: toolboxItem.icon,
-        label: I18n.t(I18nInternalNS.toolNames, toolboxItem.title || _.capitalize(tool.name)),
+        title: I18n.t(I18nInternalNS.toolNames, toolboxItem.title || _.capitalize(tool.name)),
         name: tool.name,
-        onClick: (e): void => {
+        onActivate: (): void => {
           this.toolButtonActivated(tool.name, toolboxItem.data);
         },
         description: toolboxItem.description,
@@ -293,40 +270,17 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
     };
 
     return this.toolsToBeDisplayed
-      .reduce((result, tool) => {
+      .reduce<PopoverItem[]>((result, tool) => {
         if (Array.isArray(tool.toolbox)) {
           tool.toolbox.forEach(item => {
             result.push(toPopoverItem(item, tool));
           });
-        } else {
+        } else if (tool.toolbox !== undefined)  {
           result.push(toPopoverItem(tool.toolbox, tool));
         }
 
         return result;
       }, []);
-  }
-
-  /**
-   * Validates tool's toolbox settings
-   *
-   * @param toolToolboxSettings - item to validate
-   * @param toolName - name of the tool used in console warning if item is not valid
-   */
-  private areToolboxSettingsValid(toolToolboxSettings: ToolboxConfigEntry, toolName: string): boolean {
-    /**
-     * Skip tools that don't pass 'toolbox' property
-     */
-    if (!toolToolboxSettings) {
-      return false;
-    }
-
-    if (toolToolboxSettings && !toolToolboxSettings.icon) {
-      _.log('Toolbar icon is missed. Tool %o skipped', 'warn', toolName);
-
-      return false;
-    }
-
-    return true;
   }
 
   /**
@@ -354,6 +308,26 @@ export default class Toolbox extends EventsDispatcher<ToolboxEvent> {
       on: this.api.ui.nodes.redactor,
       handler: (event: KeyboardEvent) => {
         event.preventDefault();
+
+        const currentBlockIndex = this.api.blocks.getCurrentBlockIndex();
+        const currentBlock = this.api.blocks.getBlockByIndex(currentBlockIndex);
+
+        /**
+         * Try to convert current Block to shortcut's tool
+         * If conversion is not possible, insert a new Block below
+         */
+        if (currentBlock) {
+          try {
+            this.api.blocks.convert(currentBlock.id, toolName);
+
+            window.requestAnimationFrame(() => {
+              this.api.caret.setToBlock(currentBlockIndex, 'end');
+            });
+
+            return;
+          } catch (error) {}
+        }
+
         this.insertNewBlock(toolName);
       },
     });
