@@ -24,7 +24,6 @@ declare const VERSION: string;
  * @typedef {object} ChainData
  * @property {object} data - data that will be passed to the success or fallback
  * @property {Function} function - function's that must be called asynchronously
- *
  * @interface ChainData
  */
 export interface ChainData {
@@ -38,7 +37,7 @@ export interface ChainData {
  */
 
 /**
- * Returns basic keycodes as constants
+ * Returns basic key codes as constants
  *
  * @returns {{}}
  */
@@ -178,7 +177,6 @@ export const logLabeled = _log.bind(window, true);
  * Return string representation of the object type
  *
  * @param {*} object - object to get type
- *
  * @returns {string}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -190,7 +188,6 @@ export function typeOf(object: any): string {
  * Check if passed variable is a function
  *
  * @param {*} fn - function to check
- *
  * @returns {boolean}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -202,7 +199,6 @@ export function isFunction(fn: any): fn is (...args: any[]) => any {
  * Checks if passed argument is an object
  *
  * @param {*} v - object to check
- *
  * @returns {boolean}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -214,7 +210,6 @@ export function isObject(v: any): v is object {
  * Checks if passed argument is a string
  *
  * @param {*} v - variable to check
- *
  * @returns {boolean}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -226,7 +221,6 @@ export function isString(v: any): v is string {
  * Checks if passed argument is boolean
  *
  * @param {*} v - variable to check
- *
  * @returns {boolean}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -238,7 +232,6 @@ export function isBoolean(v: any): v is boolean {
  * Checks if passed argument is number
  *
  * @param {*} v - variable to check
- *
  * @returns {boolean}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -250,7 +243,6 @@ export function isNumber(v: any): v is number {
  * Checks if passed argument is undefined
  *
  * @param {*} v - variable to check
- *
  * @returns {boolean}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -262,7 +254,6 @@ export function isUndefined(v: any): v is undefined {
  * Check if passed function is a class
  *
  * @param {Function} fn - function to check
- *
  * @returns {boolean}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -274,7 +265,6 @@ export function isClass(fn: any): boolean {
  * Checks if object is empty
  *
  * @param {object} object - object to check
- *
  * @returns {boolean}
  */
 export function isEmpty(object: object): boolean {
@@ -296,22 +286,23 @@ export function isPromise(object: any): object is Promise<any> {
   return Promise.resolve(object) === object;
 }
 
+/* eslint-disable @typescript-eslint/no-magic-numbers */
 /**
  * Returns true if passed key code is printable (a-Z, 0-9, etc) character.
  *
  * @param {number} keyCode - key code
- *
  * @returns {boolean}
  */
 export function isPrintableKey(keyCode: number): boolean {
   return (keyCode > 47 && keyCode < 58) || // number keys
-    keyCode === 32 || keyCode === 13 || // Spacebar & return key(s)
+    keyCode === 32 || keyCode === 13 || // Space bar & return key(s)
     keyCode === 229 || // processing key input for certain languages — Chinese, Japanese, etc.
     (keyCode > 64 && keyCode < 91) || // letter keys
     (keyCode > 95 && keyCode < 112) || // Numpad keys
     (keyCode > 185 && keyCode < 193) || // ;=,-./` (in order)
     (keyCode > 218 && keyCode < 223); // [\]' (in order)
 }
+/* eslint-enable @typescript-eslint/no-magic-numbers */
 
 /**
  * Fires a promise sequence asynchronously
@@ -319,8 +310,8 @@ export function isPrintableKey(keyCode: number): boolean {
  * @param {ChainData[]} chains - list or ChainData's
  * @param {Function} success - success callback
  * @param {Function} fallback - callback that fires in case of errors
- *
  * @returns {Promise}
+ * @deprecated use PromiseQueue.ts instead
  */
 export async function sequence(
   chains: ChainData[],
@@ -333,10 +324,8 @@ export async function sequence(
    * Decorator
    *
    * @param {ChainData} chainData - Chain data
-   *
    * @param {Function} successCallback - success callback
    * @param {Function} fallbackCallback - fail callback
-   *
    * @returns {Promise}
    */
   async function waitNextBlock(
@@ -370,7 +359,6 @@ export async function sequence(
  * Make array from array-like collection
  *
  * @param {ArrayLike} collection - collection to convert to array
- *
  * @returns {Array}
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -400,7 +388,6 @@ export function delay(method: (...args: any[]) => any, timeout: number) {
  * Get file extension
  *
  * @param {File} file - file
- *
  * @returns {string}
  */
 export function getFileExtension(file: File): string {
@@ -411,7 +398,6 @@ export function getFileExtension(file: File): string {
  * Check if string is MIME type
  *
  * @param {string} type - string to check
- *
  * @returns {boolean}
  */
 export function isValidMimeType(type: string): boolean {
@@ -492,6 +478,7 @@ export function throttle(func, wait, options: {leading?: boolean; trailing?: boo
 
     const remaining = wait - (now - previous);
 
+    // eslint-disable-next-line @typescript-eslint/no-this-alias
     context = this;
 
     // eslint-disable-next-line prefer-rest-params
@@ -551,7 +538,7 @@ export function getUserOS(): {[key: string]: boolean} {
     linux: false,
   };
 
-  const userOS = Object.keys(OS).find((os: string) => navigator.appVersion.toLowerCase().indexOf(os) !== -1);
+  const userOS = Object.keys(OS).find((os: string) => window.navigator.appVersion.toLowerCase().indexOf(os) !== -1);
 
   if (userOS) {
     OS[userOS] = true;
@@ -566,7 +553,6 @@ export function getUserOS(): {[key: string]: boolean} {
  * Capitalizes first letter of the string
  *
  * @param {string} text - text to capitalize
- *
  * @returns {string}
  */
 export function capitalize(text: string): string {
@@ -610,7 +596,6 @@ export function deepMerge<T extends object>(target, ...sources): T {
  * To detect touch devices more carefully, use 'touchstart' event listener
  *
  * @see http://www.stucox.com/blog/you-cant-detect-a-touchscreen/
- *
  * @returns {boolean}
  */
 export const isTouchSupported: boolean = 'ontouchstart' in document.documentElement;
@@ -674,7 +659,9 @@ export function getValidUrl(url: string): string {
  * @returns {string}
  */
 export function generateBlockId(): string {
-  return nanoid(10);
+  const idLen = 10;
+
+  return nanoid(idLen);
 }
 
 /**
@@ -690,11 +677,10 @@ export function openTab(url: string): void {
  * Returns random generated identifier
  *
  * @param {string} prefix - identifier prefix
- *
  * @returns {string}
  */
 export function generateId(prefix = ''): string {
-  // tslint:disable-next-line:no-bitwise
+  // eslint-disable-next-line @typescript-eslint/no-magic-numbers
   return `${prefix}${(Math.floor(Math.random() * 1e8)).toString(16)}`;
 }
 
@@ -761,13 +747,18 @@ export function cacheable<Target, Value, Arguments extends unknown[] = unknown[]
   }
 
   return descriptor;
-};
+}
+
+/**
+ * All screens below this width will be treated as mobile;
+ */
+export const mobileScreenBreakpoint = 650;
 
 /**
  * True if screen has mobile size
  */
 export function isMobileScreen(): boolean {
-  return window.matchMedia('(max-width: 650px)').matches;
+  return window.matchMedia(`(max-width: ${mobileScreenBreakpoint}px)`).matches;
 }
 
 /**
