@@ -1,71 +1,81 @@
-import UI from '../components/modules/ui';
-import BlockEvents from '../components/modules/blockEvents';
-import Toolbar from '../components/modules/toolbar/index';
-import InlineToolbar from '../components/modules/toolbar/inline';
-import BlockSettings from '../components/modules/toolbar/blockSettings';
-import Paste from '../components/modules/paste';
-import DragNDrop from '../components/modules/dragNDrop';
-import Renderer from '../components/modules/renderer';
-import Tools from '../components/modules/tools';
-import API from '../components/modules/api/index';
-import Caret from '../components/modules/caret';
-import BlockManager from '../components/modules/blockManager';
+/** ./api */
 import BlocksAPI from '../components/modules/api/blocks';
 import CaretAPI from '../components/modules/api/caret';
 import EventsAPI from '../components/modules/api/events';
-import ListenersAPI from '../components/modules/api/listeners';
-import SanitizerAPI from '../components/modules/api/sanitizer';
-import ToolbarAPI from '../components/modules/api/toolbar';
-import StylesAPI from '../components/modules/api/styles';
-import SelectionAPI from '../components/modules/api/selection';
-import NotifierAPI from '../components/modules/api/notifier';
-import SaverAPI from '../components/modules/api/saver';
-import Saver from '../components/modules/saver';
-import BlockSelection from '../components/modules/blockSelection';
-import RectangleSelection from '../components/modules/RectangleSelection';
-import InlineToolbarAPI from '../components/modules/api/inlineToolbar';
-import CrossBlockSelection from '../components/modules/crossBlockSelection';
-import ConversionToolbar from '../components/modules/toolbar/conversion';
-import TooltipAPI from '../components/modules/api/tooltip';
-import ReadOnly from '../components/modules/readonly';
-import ReadOnlyAPI from '../components/modules/api/readonly';
 import I18nAPI from '../components/modules/api/i18n';
+import API from '../components/modules/api/index';
+import InlineToolbarAPI from '../components/modules/api/inlineToolbar';
+import ListenersAPI from '../components/modules/api/listeners';
+import NotifierAPI from '../components/modules/api/notifier';
+import ReadOnlyAPI from '../components/modules/api/readonly';
+import SanitizerAPI from '../components/modules/api/sanitizer';
+import SaverAPI from '../components/modules/api/saver';
+import SelectionAPI from '../components/modules/api/selection';
+import StylesAPI from '../components/modules/api/styles';
+import ToolbarAPI from '../components/modules/api/toolbar';
+import TooltipAPI from '../components/modules/api/tooltip';
 import UiAPI from '../components/modules/api/ui';
+
+/** ./toolbar */
+import BlockSettings from '../components/modules/toolbar/blockSettings';
+import ConversionToolbar from '../components/modules/toolbar/conversion';
+import Toolbar from '../components/modules/toolbar/index';
+import InlineToolbar from '../components/modules/toolbar/inline';
+
+/** . */
+import BlockEvents from '../components/modules/blockEvents';
+import BlockManager from '../components/modules/blockManager';
+import BlockSelection from '../components/modules/blockSelection';
+import Caret from '../components/modules/caret';
+import CrossBlockSelection from '../components/modules/crossBlockSelection';
+import DragNDrop from '../components/modules/dragNDrop';
 import ModificationsObserver from '../components/modules/modificationsObserver';
+import Paste from '../components/modules/paste';
+import ReadOnly from '../components/modules/readonly';
+import RectangleSelection from '../components/modules/rectangleSelection';
+import Renderer from '../components/modules/renderer';
+import Saver from '../components/modules/saver';
+import Tools from '../components/modules/tools';
+import UI from '../components/modules/ui';
 
 export interface EditorModules {
-  UI: UI;
-  BlockEvents: BlockEvents;
-  BlockSelection: BlockSelection;
-  RectangleSelection: RectangleSelection;
-  Toolbar: Toolbar;
-  InlineToolbar: InlineToolbar;
-  BlockSettings: BlockSettings;
-  ConversionToolbar: ConversionToolbar;
-  Paste: Paste;
-  DragNDrop: DragNDrop;
-  Renderer: Renderer;
-  Tools: Tools;
-  API: API;
-  Caret: Caret;
-  Saver: Saver;
-  BlockManager: BlockManager;
-  BlocksAPI: BlocksAPI;
-  CaretAPI: CaretAPI;
-  EventsAPI: EventsAPI;
-  ListenersAPI: ListenersAPI;
-  SanitizerAPI: SanitizerAPI;
-  SaverAPI: SaverAPI;
-  SelectionAPI: SelectionAPI;
-  StylesAPI: StylesAPI;
-  ToolbarAPI: ToolbarAPI;
-  InlineToolbarAPI: InlineToolbarAPI;
-  CrossBlockSelection: CrossBlockSelection;
-  NotifierAPI: NotifierAPI;
-  TooltipAPI: TooltipAPI;
-  ReadOnly: ReadOnly;
-  ReadOnlyAPI: ReadOnlyAPI;
-  I18nAPI: I18nAPI;
-  UiAPI: UiAPI;
-  ModificationsObserver: ModificationsObserver;
+  // API Modules
+  BlocksAPI: BlocksAPI,
+  CaretAPI: CaretAPI,
+  EventsAPI: EventsAPI,
+  I18nAPI: I18nAPI,
+  API: API,
+  InlineToolbarAPI: InlineToolbarAPI,
+  ListenersAPI: ListenersAPI,
+  NotifierAPI: NotifierAPI,
+  ReadOnlyAPI: ReadOnlyAPI,
+  SanitizerAPI: SanitizerAPI,
+  SaverAPI: SaverAPI,
+  SelectionAPI: SelectionAPI,
+  StylesAPI: StylesAPI,
+  ToolbarAPI: ToolbarAPI,
+  TooltipAPI: TooltipAPI,
+  UiAPI: UiAPI,
+
+  // Toolbar Modules
+  BlockSettings: BlockSettings,
+  ConversionToolbar: ConversionToolbar,
+  Toolbar: Toolbar,
+  InlineToolbar: InlineToolbar,
+
+  // Modules
+  BlockEvents: BlockEvents,
+  BlockManager: BlockManager,
+  BlockSelection: BlockSelection,
+  Caret: Caret,
+  CrossBlockSelection: CrossBlockSelection,
+  DragNDrop: DragNDrop,
+  ModificationsObserver: ModificationsObserver,
+  Paste: Paste,
+  ReadOnly: ReadOnly,
+  RectangleSelection: RectangleSelection,
+  Renderer: Renderer,
+  Saver: Saver,
+  Tools: Tools,
+  UI: UI,
 }
