@@ -61,7 +61,7 @@ Cypress.Commands.add('paste', {
 
   subject[0].dispatchEvent(pasteEvent);
 
-  return subject;
+  cy.wait(200); // wait a little since some tools (paragraph) could have async hydration
 });
 
 /**
@@ -70,7 +70,7 @@ Cypress.Commands.add('paste', {
  * Usage:
  * cy.get('div').copy().then(data => {})
  */
-Cypress.Commands.add('copy', { prevSubject: true }, async (subject) => {
+Cypress.Commands.add('copy', { prevSubject: true }, (subject) => {
   const clipboardData: {[type: string]: any} = {};
 
   const copyEvent = Object.assign(new Event('copy', {
@@ -79,7 +79,6 @@ Cypress.Commands.add('copy', { prevSubject: true }, async (subject) => {
   }), {
     clipboardData: {
       setData: (type: string, data: any): void => {
-        console.log(type, data);
         clipboardData[type] = data;
       },
     },
@@ -87,7 +86,7 @@ Cypress.Commands.add('copy', { prevSubject: true }, async (subject) => {
 
   subject[0].dispatchEvent(copyEvent);
 
-  return clipboardData;
+  return cy.wrap(clipboardData);
 });
 
 /**
@@ -96,7 +95,7 @@ Cypress.Commands.add('copy', { prevSubject: true }, async (subject) => {
  * Usage:
  * cy.get('div').cut().then(data => {})
  */
-Cypress.Commands.add('cut', { prevSubject: true }, async (subject) => {
+Cypress.Commands.add('cut', { prevSubject: true }, (subject) => {
   const clipboardData: {[type: string]: any} = {};
 
   const copyEvent = Object.assign(new Event('cut', {
@@ -105,7 +104,6 @@ Cypress.Commands.add('cut', { prevSubject: true }, async (subject) => {
   }), {
     clipboardData: {
       setData: (type: string, data: any): void => {
-        console.log(type, data);
         clipboardData[type] = data;
       },
     },
@@ -113,7 +111,7 @@ Cypress.Commands.add('cut', { prevSubject: true }, async (subject) => {
 
   subject[0].dispatchEvent(copyEvent);
 
-  return clipboardData;
+  return cy.wrap(clipboardData);
 });
 
 /**
@@ -121,8 +119,39 @@ Cypress.Commands.add('cut', { prevSubject: true }, async (subject) => {
  *
  * @param data — data to render
  */
-Cypress.Commands.add('render', { prevSubject: true }, async (subject: EditorJS, data: OutputData): Promise<EditorJS> => {
-  await subject.render(data);
+Cypress.Commands.add('render', { prevSubject: true }, (subject: EditorJS, data: OutputData) => {
+  return cy.wrap(subject.render(data))
+    .then(() => {
+      return cy.wrap(subject);
+    });
+});
 
-  return subject;
+
+/**
+ * Select passed text in element
+ * Note. Previous subject should have 'textNode' as firstChild
+ *
+ * Usage
+ * cy.get('[data-cy=editorjs]')
+ *  .find('.ce-paragraph')
+ *  .selectText('block te')
+ *
+ * @param text - text to select
+ */
+Cypress.Commands.add('selectText', {
+  prevSubject: true,
+}, (subject, text: string) => {
+  const el = subject[0];
+  const document = el.ownerDocument;
+  const range = document.createRange();
+  const textNode = el.firstChild;
+  const selectionPositionStart = textNode.textContent.indexOf(text);
+  const selectionPositionEnd = selectionPositionStart + text.length;
+
+  range.setStart(textNode, selectionPositionStart);
+  range.setEnd(textNode, selectionPositionEnd);
+  document.getSelection().removeAllRanges();
+  document.getSelection().addRange(range);
+
+  return cy.wrap(subject);
 });
