@@ -1,4 +1,4 @@
-import {OutputData} from '../data-formats/output-data';
+import {OutputBlockData, OutputData} from '../data-formats/output-data';
 import {BlockToolData, ToolConfig} from '../tools';
 import {BlockAPI} from './block';
 
@@ -9,7 +9,7 @@ export interface Blocks {
   /**
    * Remove all blocks from Editor zone
    */
-  clear(): void;
+  clear(): Promise<void>;
 
   /**
    * Render passed data
@@ -52,7 +52,7 @@ export interface Blocks {
    * Returns Block API object by passed Block index
    * @param {number} index
    */
-  getBlockByIndex(index: number): BlockAPI | void;
+  getBlockByIndex(index: number): BlockAPI | undefined;
 
   /**
    * Returns Block API object by passed Block id
@@ -126,6 +126,7 @@ export interface Blocks {
    * @param {number?} index — index where to insert new Block
    * @param {boolean?} needToFocus - flag to focus inserted Block
    * @param {boolean?} replace - should the existed Block on that index be replaced or not
+   * @param {string} id — An optional id for the new block. If omitted then the new id will be generated
    */
   insert(
     type?: string,
@@ -134,7 +135,16 @@ export interface Blocks {
     index?: number,
     needToFocus?: boolean,
     replace?: boolean,
+    id?: string,
   ): BlockAPI;
+
+  /**
+   * Inserts several Blocks to specified index
+   */
+  insertMany(
+    blocks: OutputBlockData[],
+    index?: number,
+  ): BlockAPI[];
 
 
   /**
@@ -148,7 +158,18 @@ export interface Blocks {
    * Updates block data by id
    *
    * @param id - id of the block to update
-   * @param data - the new data
+   * @param data - the new data. Can be partial.
    */
-  update(id: string, data: BlockToolData): void;
+  update(id: string, data: Partial<BlockToolData>): Promise<BlockAPI>;
+
+  /**
+   * Converts block to another type. Both blocks should provide the conversionConfig.
+   *
+   * @param id - id of the existed block to convert. Should provide 'conversionConfig.export' method
+   * @param newType - new block type. Should provide 'conversionConfig.import' method
+   * @param dataOverrides - optional data overrides for the new block
+   *
+   * @throws Error if conversion is not possible
+   */
+  convert(id: string, newType: string, dataOverrides?: BlockToolData): void;
 }
