@@ -31,6 +31,11 @@ import {
 } from './api';
 
 import { OutputData } from './data-formats';
+import { BlockMutationEvent, BlockMutationEventMap, BlockMutationType } from './events/block';
+import { BlockAddedMutationType, BlockAddedEvent } from './events/block/BlockAdded';
+import { BlockChangedMutationType, BlockChangedEvent } from './events/block/BlockChanged';
+import { BlockMovedMutationType, BlockMovedEvent } from './events/block/BlockMoved';
+import { BlockRemovedMutationType, BlockRemovedEvent } from './events/block/BlockRemoved';
 
 /**
  * Interfaces used for development
@@ -64,6 +69,7 @@ export {BlockTune, BlockTuneConstructable} from './block-tunes';
 export {
   EditorConfig,
   SanitizerConfig,
+  SanitizerRule,
   PasteConfig,
   LogLevels,
   ConversionConfig,
@@ -71,9 +77,26 @@ export {
   Dictionary,
   DictValue,
   I18nConfig,
+  PopoverItem,
+  PopoverItemWithConfirmation,
+  PopoverItemWithoutConfirmation
 } from './configs';
-export {OutputData, OutputBlockData} from './data-formats/output-data';
+export { OutputData, OutputBlockData} from './data-formats/output-data';
+export { BlockId } from './data-formats/block-id';
 export { BlockAPI } from './api'
+export {
+  BlockMutationType,
+  BlockMutationEvent,
+  BlockMutationEventMap,
+  BlockAddedMutationType,
+  BlockAddedEvent,
+  BlockRemovedMutationType,
+  BlockRemovedEvent,
+  BlockMovedMutationType,
+  BlockMovedEvent,
+  BlockChangedMutationType,
+  BlockChangedEvent,
+}
 
 /**
  * We have a namespace API {@link ./api/index.d.ts} (APIMethods) but we can not use it as interface
