@@ -5,6 +5,7 @@ import { ToolConfig } from './tool-config';
 import { API, BlockAPI, ToolboxConfig } from '../index';
 import { PasteEvent } from './paste-events';
 import { MoveEvent } from './hook-events';
+import { TunesMenuConfig } from './tool-settings';
 
 /**
  * Describe Block Tool object
@@ -25,9 +26,8 @@ export interface BlockTool extends BaseTool {
 
   /**
    * Create Block's settings block
-   * @return {HTMLElement}
    */
-  renderSettings?(): HTMLElement;
+  renderSettings?(): HTMLElement | TunesMenuConfig;
 
   /**
    * Validate Block's data
